Stop double scroll issue
