+++ web-embeds/public/index.html
@@ -1,21 +1,19 @@
 <!doctype html>
 <html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="theme-color" content="#000000" />
-    <meta
-      name="description"
-      content="Web embed created using The Apollos Project"
-    />
-    <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
-    <!--
+
+<head>
+  <meta charset="utf-8" />
+  <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
+  <meta name="viewport" content="width=device-width, initial-scale=1" />
+  <meta name="theme-color" content="#000000" />
+  <meta name="description" content="Web embed created using The Apollos Project" />
+  <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
+  <!--
       manifest.json provides metadata used when your web app is installed on a
       user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
     -->
-    <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
-    <!--
+  <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
+  <!--
       Notice the use of %PUBLIC_URL% in the tags above.
       It will be replaced with the URL of the `public` folder during the build.
       Only files inside the `public` folder can be referenced from the HTML.
@@ -24,30 +22,31 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>Apollos Web Embeds</title>
-    <style>
-      body {
-        font-family:
-          "Inter",
-          -apple-system,
-          BlinkMacSystemFont,
-          Segoe UI,
-          Roboto,
-          Oxygen,
-          Ubuntu,
-          Cantarell,
-          Fira Sans,
-          Droid Sans,
-          Helvetica Neue,
-          sans-serif;
-      }
-    </style>
-  </head>
+  <title>Apollos Web Embeds</title>
+  <style>
+    body {
+      font-family:
+        "Inter",
+        -apple-system,
+        BlinkMacSystemFont,
+        Segoe UI,
+        Roboto,
+        Oxygen,
+        Ubuntu,
+        Cantarell,
+        Fira Sans,
+        Droid Sans,
+        Helvetica Neue,
+        sans-serif;
+      background-color: aqua;
+    }
+  </style>
+</head>
 
-  <body>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
-    <!-- This is an example of how to use the apollos widgets -->
-    <!-- <div
+<body>
+  <noscript>You need to enable JavaScript to run this app.</noscript>
+  <!-- This is an example of how to use the apollos widgets -->
+  <!-- <div
       data-church="cedar_creek"
       data-type="FeatureFeed"
       data-feature-feed="FeatureFeed:5aae43e6-3526-4cd2-8dfe-771d2ce8a333"
@@ -57,7 +56,7 @@
       style="max-width: 1180px; padding: 40px; margin: auto; margin-top: 20px"
     ></div> -->
 
-    <!-- <div
+  <!-- <div
       data-type="FeatureFeed"
       data-church="liquid_church"
       data-feature-feed="FeatureFeed:d292fd8c-5b9f-4cd7-a2fe-9d7d0914c34f"
@@ -65,23 +64,20 @@
       data-modal="true"
       class="apollos-widget"
     ></div> -->
-    <!-- 
-    <div
-      data-church="cedar_creek"
-      data-type="Search"
-      data-search-feed="FeatureFeed:33b34f30-e57c-4c23-9f28-24d31959a3f4"
-      data-modal="true"
-      class="apollos-widget"
-      style="max-width: 1180px; padding: 40px; margin: auto; margin-top: 20px"
-    ></div> -->
-    <div
+  <div data-church="cedar_creek" data-type="Search" data-search-feed="FeatureFeed:33b34f30-e57c-4c23-9f28-24d31959a3f4"
+    data-modal="true" class="apollos-widget" style="max-width: 1180px; padding: 40px; margin: auto; margin-top: 20px">
+  </div>
+  <div style="height: 1000px; border: 1px solid red; width: 50vw;">
+
+  </div>
+  <!-- <div
       data-church="bayside"
       data-type="Auth"
       data-modal="true"
       class="apollos-widget"
       style="max-width: 1180px; padding: 40px; margin: auto; margin-top: 20px"
-    ></div>
-    <!--
+    ></div> -->
+  <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
 
@@ -91,5 +87,6 @@
       To begin the development, run `npm start` or `yarn start`.
       To create a production bundle, use `npm run build` or `yarn build`.
     -->
-  </body>
+</body>
+
 </html>
